refactor: ソロラウンドカットインコンポーネントのリファクタリング。
